world_bank_data/requirements.txt
requests>=2.28.0
pandas>=1.5.0
neo4j>=5.0.0
python-dotenv>=1.0.0 